Practical Assignment 1
The assignment should be completed on vSphere, but you can practice on VirtualBox.
The Linux distribution should not matter as long as you use the correct package manager. Most of the assignment is based on shell/file commands.
Recommend fully installing and updating your VMs, then creating snapshots in vSphere so you can always revert back if you need to.
You will need at least one virtual machine with a Desktop version of Linux for the GUI, in order to use a browser to download and upload files.
You will also need at least one virtual machine with a Server version of Linux for the CLI, as part of the assignment will require you to transfer files between Linux machines using rsync or scp/sftp.